data-ingestion/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.0
beautifulsoup4==4.14.3
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==44.0.3
cymem==2.0.13
distlib==0.4.0
distro==1.9.0
einops==0.8.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
fastembed==0.7.4
filelock==3.20.1
filetype==1.2.0
FlashRank==0.2.10
flatbuffers==25.12.19
fsspec==2025.12.0
ftfy==6.3.1
google-auth==2.45.0
google-genai==1.56.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
loguru==0.7.3
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.1
MarkupSafe==3.0.3
mmh3==5.2.0
mpmath==1.3.0
murmurhash==1.0.15
neo4j==6.0.3
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.0
onnxruntime==1.23.2
openai==1.109.1
opencv-python-headless==4.11.0.86
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdftext==0.6.3
phonenumbers==9.0.21
pillow==10.4.0
platformdirs==4.5.1
portalocker==3.2.0
pre_commit==4.5.1
preshed==3.0.12
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
protobuf==6.33.2
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
RapidFuzz==3.14.3
regex==2024.11.6
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
surya-ocr==0.17.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
virtualenv==20.35.4
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
zstandard==0.25.0
